added functions to work viwe archive
archive_get_files-The function allows you to get information about the files contained in the archive, including their number, size and other characteristics.
archive_save-This function is responsible for creating a new archive or updating an existing one by adding new files to it.
archive_extract-this function extracts selected files from the archive using decompression using Huffman encoding. It opens files for writing along the specified paths and writes decompressed data to them. If the file is not included in the selected list, then it is skipped.
archive_remove_files-This function deletes selected files from the archive by shifting the remaining files in the file stream to fill in the gaps left after deletion. It also updates the file counter in the archive and truncates the file to the new size. This allows you to maintain the integrity of the archive after deleting files.
archive_free-This function is necessary to ensure the integrity of the archive and clean up the resources associated with it in case of errors or shutdown of the archive.