Update main.c
The crate and extract functions have been edited, the read_and_dev_block, read_and_dec_file read_and_de_file functions
have been
added - reads a compressed file that has been archived using Huffman encoding, and decodes it, writing the decompressed data to the specified file. Decoding takes place in blocks, according to the size of the buffer that is passed to the function. In addition, the function updates information about the current decompression progress in the archive structure.
read_and_dev_block - reads a block of data from a file, decodes it, and writes the result back to another file. The basic block size and additional information such as padding and padding are used to correctly read and decode data. After decoding, the data is written to the final buffer and then written to the output file, taking into account the correctness of data recording after decoding.