Update main.c
The content for the add, extract, delete functions was added to the file for user interaction, the content of the create function was changed and improved, checks for the existence of files and directories entered by the user were added, as well as restrictions on entering commands (non-existent commands are not entered). In addition, the compatibility of the code with various development environments has been improved. The "help" function was created and filled in, and the "test" function was added to verify the existence of files.